Flytt visitor-klasse til egen fil
